Made useeffect into a separate hook
